Narrow Rational from i128 to i64 components
Rational is a bounded exact type; its `(numer, denom)` pair now uses i64
rather than i128. The trap ceiling drops to ~18 significant digits /
denominators up to ~9.2e18 — plenty for decimal literals, money, and
modest ratios, and it makes the representation lighter across the board.

Arithmetic runs in i128 intermediates (widen operands, checked mul/add,
reduce, narrow to i64), so a cross-multiply never wraps before reduction;
only a reduced result that still exceeds i64 traps. `reduce_to_i64` is the
single narrowing funnel (runtime), mirrored by `reduce_rational` at
compile time.

Cells shrink 32 -> 16 bytes (num @ 0, den @ 8). A welcome side effect: the
i128 alignment machinery deletes itself — the LLVM `align 16`/`align 8`
annotations are gone (i64 is naturally 8-aligned, matching the record
buffer), and Cranelift drops the `iconcat`-based i128 constant helper (a
plain `iconst` holds a full i64) and its 16-aligned stack slots become 8.

Behavior is otherwise unchanged: equality, `/`->Rational, `div`, rational
arithmetic, conversions, cells, and SQL pushdown all still pass on both
backends (176 e2e). Storage stays canonical TEXT "n/d".

+++ crates/coddl-procir/src/ir.rs
@@ -619,9 +619,9 @@ pub enum Const {
     /// equality is a proper (reflexive) equality. `double` at the ABI.
     Approximate(u64),
     /// A bounded exact `Rational` as a **reduced** `(numer, denom)` pair
-    /// (`gcd(|n|,d) = 1`, `d > 0`, `0 = (0,1)`). Two `i128`s — a compound
-    /// value at the ABI (like `Text`'s `(ptr, len)`), 32-byte cell.
-    Rational(i128, i128),
+    /// (`gcd(|n|,d) = 1`, `d > 0`, `0 = (0,1)`). Two `i64`s — a compound
+    /// value at the ABI (like `Text`'s `(ptr, len)`), 16-byte cell.
+    Rational(i64, i64),
     /// `true` / `false` — Boolean literal value.
     Boolean(bool),
     /// The `Tuple {}` value — produced where the source had `{}` or
