+++ crates/coddl-sqlemit/src/lib.rs
@@ -92,7 +92,7 @@ pub enum Value {
     Approximate(u64),
     /// A bounded `Rational` as its reduced `(numer, denom)` pair. Binds/reads
     /// as canonical SQL `TEXT "n/d"`.
-    Rational(i128, i128),
+    Rational(i64, i64),
     Boolean(bool),
 }
